+++ src/main/resources/db/migration/20180525070315483__add_job_title_and_phone_number_columns_to_user.sql
@@ -1,0 +1,3 @@
+ALTER TABLE referencedata.users
+  ADD COLUMN jobTitle CHARACTER VARYING(255),
+  ADD COLUMN phoneNumber CHARACTER VARYING(255);